refactoring: separated rendering utilities as render_utils.py
