+++ render_images.py
@@ -6,24 +6,9 @@
 # of patent rights can be found in the PATENTS file in the same directory.
 
 from __future__ import print_function
-import math, sys, random, argparse, json, os, tempfile
+import sys, argparse, json, os
 from datetime import datetime as dt
-from collections import Counter
-from copy import deepcopy as copy
-import numpy as np
   
-"""
-Renders random scenes using Blender, each with with a random number of objects;
-each object has a random size, position, color, and shape. Objects will be
-nonintersecting but may partially occlude each other. Output images will be
-written to disk as PNGs, and we will also write a JSON file for each image with
-ground-truth scene information.
-
-This file expects to be run from Blender like this:
-
-blender --background --python render_images.py -- [arguments to this script]
-"""
-
 INSIDE_BLENDER = True
 try:
   import bpy, bpy_extras
@@ -35,6 +20,7 @@ if INSIDE_BLENDER:
     import utils
     import blocks
     from blocks import State, Unstackable
+    from render_utils import render_scene
   except ImportError as e:
     print("\nERROR")
     print("Running render_images.py from Blender and cannot import utils.py.") 
@@ -184,288 +170,6 @@ def main(args):
       except Unstackable as e:
         print(e)
         pass
-
-
-def render_scene(args,
-    output_image='render.png',
-    output_scene='render_json',
-    output_blendfile=None,
-    objects=[],
-    **kwargs
-  ):
-
-  # Load the main blendfile
-  bpy.ops.wm.open_mainfile(filepath=args.base_scene_blendfile)
-
-  # Load materials
-  utils.load_materials(args.material_dir)
-
-  # Set render arguments so we can get pixel coordinates later.
-  # We use functionality specific to the CYCLES renderer so BLENDER_RENDER
-  # cannot be used.
-  render_args = bpy.context.scene.render
-  render_args.engine = "CYCLES"
-  render_args.filepath = output_image
-  render_args.resolution_x = args.width
-  render_args.resolution_y = args.height
-  render_args.resolution_percentage = 100
-  render_args.tile_x = args.render_tile_size
-  render_args.tile_y = args.render_tile_size
-  if args.use_gpu == 1:
-    # Blender changed the API for enabling CUDA at some point
-    if bpy.app.version < (2, 78, 0):
-      bpy.context.user_preferences.system.compute_device_type = 'CUDA'
-      bpy.context.user_preferences.system.compute_device = 'CUDA_0'
-    else:
-      cycles_prefs = bpy.context.user_preferences.addons['cycles'].preferences
-      cycles_prefs.compute_device_type = 'CUDA'
-
-  # Some CYCLES-specific stuff
-  bpy.data.worlds['World'].cycles.sample_as_light = True
-  bpy.context.scene.cycles.blur_glossy = 2.0
-  bpy.context.scene.cycles.samples = args.render_num_samples
-  bpy.context.scene.cycles.transparent_min_bounces = args.render_min_bounces
-  bpy.context.scene.cycles.transparent_max_bounces = args.render_max_bounces
-  if args.use_gpu == 1:
-    bpy.context.scene.cycles.device = 'GPU'
-
-  # This will give ground-truth information about the scene and its objects
-  scene_struct = {
-      'image_filename': os.path.basename(output_image),
-      'objects': [],
-      'directions': {},
-  }
-  scene_struct.update(kwargs)
-
-  if bpy.app.version < (2, 80, 0):
-    bpy.ops.mesh.primitive_plane_add(radius=5)
-  else:
-    bpy.ops.mesh.primitive_plane_add(size=5)
-
-  plane = bpy.context.object
-
-  def rand(L):
-    return 2.0 * L * (random.random() - 0.5)
-
-  # Add random jitter to camera position
-  if args.camera_jitter > 0:
-    for i in range(3):
-      bpy.data.objects['Camera'].location[i] += rand(args.camera_jitter)
-
-  # Figure out the left, up, and behind directions along the plane and record
-  # them in the scene structure
-  camera = bpy.data.objects['Camera']
-  plane_normal = plane.data.vertices[0].normal
-  if bpy.app.version < (2, 80, 0):
-    cam_behind = camera.matrix_world.to_quaternion() * Vector((0, 0, -1))
-    cam_left = camera.matrix_world.to_quaternion() * Vector((-1, 0, 0))
-    cam_up = camera.matrix_world.to_quaternion() * Vector((0, 1, 0))
-  else:
-    cam_behind = camera.matrix_world.to_quaternion() @ Vector((0, 0, -1))
-    cam_left = camera.matrix_world.to_quaternion() @ Vector((-1, 0, 0))
-    cam_up = camera.matrix_world.to_quaternion() @ Vector((0, 1, 0))
-  plane_behind = (cam_behind - cam_behind.project(plane_normal)).normalized()
-  plane_left = (cam_left - cam_left.project(plane_normal)).normalized()
-  plane_up = cam_up.project(plane_normal).normalized()
-
-  # Delete the plane; we only used it for normals anyway. The base scene file
-  # contains the actual ground plane.
-  utils.delete_object(plane)
-
-  # Save all six axis-aligned directions in the scene struct
-  scene_struct['directions']['behind'] = tuple(plane_behind)
-  scene_struct['directions']['front'] = tuple(-plane_behind)
-  scene_struct['directions']['left'] = tuple(plane_left)
-  scene_struct['directions']['right'] = tuple(-plane_left)
-  scene_struct['directions']['above'] = tuple(plane_up)
-  scene_struct['directions']['below'] = tuple(-plane_up)
-
-  # Add random jitter to lamp positions
-  if args.key_light_jitter > 0:
-    for i in range(3):
-      bpy.data.objects['Lamp_Key'].location[i] += rand(args.key_light_jitter)
-  if args.back_light_jitter > 0:
-    for i in range(3):
-      bpy.data.objects['Lamp_Back'].location[i] += rand(args.back_light_jitter)
-  if args.fill_light_jitter > 0:
-    for i in range(3):
-      bpy.data.objects['Lamp_Fill'].location[i] += rand(args.fill_light_jitter)
-
-  # Now make some random objects
-  blender_objects = add_objects(scene_struct, camera, objects)
-
-  # Render the scene and dump the scene data structure
-  scene_struct['objects'] = objects
-  scene_struct['relationships'] = compute_all_relationships(scene_struct)
-  while True:
-    try:
-      bpy.ops.render.render(write_still=True)
-      break
-    except Exception as e:
-      print(e)
-
-  with open(output_scene, 'w') as f:
-    json.dump(scene_struct, f, indent=2)
-
-  if output_blendfile is not None:
-    bpy.ops.wm.save_as_mainfile(filepath=output_blendfile)
-
-
-def add_objects(scene_struct, camera, objects):
-  """
-  Add objects to the current blender scene
-  """
-  blender_objects = []
-  for obj in objects:
-    
-    # Actually add the object to the scene
-    utils.add_object(args.shape_dir,
-                     obj["shape"],
-                     obj["size"],
-                     obj["location"],
-                     theta=obj["rotation"])
-    bobj = bpy.context.object
-    blender_objects.append(bobj)
-    utils.add_material(obj["material"], Color=obj["color"])
-    obj["pixel_coords"] = utils.get_camera_coords(camera, bobj.location)
-
-    loc = np.array(bobj.location)
-    dim = np.array(bobj.dimensions)
-    half = dim / 2
-    corners = []
-    corners.append(loc + half * [1,1,1])
-    corners.append(loc + half * [1,1,-1])
-    corners.append(loc + half * [1,-1,1])
-    corners.append(loc + half * [1,-1,-1])
-    corners.append(loc + half * [-1,1,1])
-    corners.append(loc + half * [-1,1,-1])
-    corners.append(loc + half * [-1,-1,1])
-    corners.append(loc + half * [-1,-1,-1])
-
-    import mathutils
-    corners_camera_coords = np.array([ utils.get_camera_coords(camera, mathutils.Vector(tuple(corner)))
-                                       for corner in corners ])
-    xmax = np.amax(corners_camera_coords[:,0])
-    ymax = np.amax(corners_camera_coords[:,1])
-    xmin = np.amin(corners_camera_coords[:,0])
-    ymin = np.amin(corners_camera_coords[:,1])
-    obj["bbox"] = (float(xmin), float(ymin), float(xmax), float(ymax))
-  return blender_objects
-
-
-def compute_all_relationships(scene_struct, eps=0.2):
-  """
-  Computes relationships between all pairs of objects in the scene.
-  
-  Returns a dictionary mapping string relationship names to lists of lists of
-  integers, where output[rel][i] gives a list of object indices that have the
-  relationship rel with object i. For example if j is in output['left'][i] then
-  object j is left of object i.
-  """
-  all_relationships = {}
-  for name, direction_vec in scene_struct['directions'].items():
-    if name == 'above' or name == 'below': continue
-    all_relationships[name] = []
-    for i, obj1 in enumerate(scene_struct['objects']):
-      coords1 = obj1['location']
-      related = set()
-      for j, obj2 in enumerate(scene_struct['objects']):
-        if obj1 == obj2: continue
-        coords2 = obj2['location']
-        diff = [coords2[k] - coords1[k] for k in [0, 1, 2]]
-        dot = sum(diff[k] * direction_vec[k] for k in [0, 1, 2])
-        if dot > eps:
-          related.add(j)
-      all_relationships[name].append(sorted(list(related)))
-  return all_relationships
-
-
-def check_visibility(blender_objects, min_pixels_per_object):
-  """
-  Check whether all objects in the scene have some minimum number of visible
-  pixels; to accomplish this we assign random (but distinct) colors to all
-  objects, and render using no lighting or shading or antialiasing; this
-  ensures that each object is just a solid uniform color. We can then count
-  the number of pixels of each color in the output image to check the visibility
-  of each object.
-
-  Returns True if all objects are visible and False otherwise.
-  """
-  f, path = tempfile.mkstemp(suffix='.png')
-  object_colors = render_shadeless(blender_objects, path=path)
-  img = bpy.data.images.load(path)
-  p = list(img.pixels)
-  color_count = Counter((p[i], p[i+1], p[i+2], p[i+3])
-                        for i in range(0, len(p), 4))
-  os.remove(path)
-  if len(color_count) != len(blender_objects) + 1:
-    return False
-  for _, count in color_count.most_common():
-    if count < min_pixels_per_object:
-      return False
-  return True
-
-
-def render_shadeless(blender_objects, path='flat.png'):
-  """
-  Render a version of the scene with shading disabled and unique materials
-  assigned to all objects, and return a set of all colors that should be in the
-  rendered image. The image itself is written to path. This is used to ensure
-  that all objects will be visible in the final rendered scene.
-  """
-  render_args = bpy.context.scene.render
-
-  # Cache the render args we are about to clobber
-  old_filepath = render_args.filepath
-  old_engine = render_args.engine
-  old_use_antialiasing = render_args.use_antialiasing
-
-  # Override some render settings to have flat shading
-  render_args.filepath = path
-  render_args.engine = 'BLENDER_RENDER'
-  render_args.use_antialiasing = False
-
-  # Move the lights and ground to layer 2 so they don't render
-  utils.set_layer(bpy.data.objects['Lamp_Key'], 2)
-  utils.set_layer(bpy.data.objects['Lamp_Fill'], 2)
-  utils.set_layer(bpy.data.objects['Lamp_Back'], 2)
-  utils.set_layer(bpy.data.objects['Ground'], 2)
-
-  # Add random shadeless materials to all objects
-  object_colors = set()
-  old_materials = []
-  for i, obj in enumerate(blender_objects):
-    old_materials.append(obj.data.materials[0])
-    bpy.ops.material.new()
-    mat = bpy.data.materials['Material']
-    mat.name = 'Material_%d' % i
-    while True:
-      r, g, b = [random.random() for _ in range(3)]
-      if (r, g, b) not in object_colors: break
-    object_colors.add((r, g, b))
-    mat.diffuse_color = [r, g, b]
-    mat.use_shadeless = True
-    obj.data.materials[0] = mat
-
-  # Render the scene
-  bpy.ops.render.render(write_still=True)
-
-  # Undo the above; first restore the materials to objects
-  for mat, obj in zip(old_materials, blender_objects):
-    obj.data.materials[0] = mat
-
-  # Move the lights and ground back to layer 0
-  utils.set_layer(bpy.data.objects['Lamp_Key'], 0)
-  utils.set_layer(bpy.data.objects['Lamp_Fill'], 0)
-  utils.set_layer(bpy.data.objects['Lamp_Back'], 0)
-  utils.set_layer(bpy.data.objects['Ground'], 0)
-
-  # Set the render settings back to what they were
-  render_args.filepath = old_filepath
-  render_args.engine = old_engine
-  render_args.use_antialiasing = old_use_antialiasing
-
-  return object_colors
 
 
 if __name__ == '__main__':
